requirements.txt
flask==3.0.0
google-cloud-translate==3.15.0
python-dotenv==1.0.0
anthropic==0.39.0
